started etch-a-sketch project
there is now an html file, a css file, and a js file
in the html file there is one div with an id of grid.
within the js file there is a function that creates a grid of divs
within the grid div.
in the css file i have set up the actual grid using calc(100%/16) for
width and padding-bottom to make the divs square. i did this for responsiveness.