the large canvas button now works
the large option now works, but there is a small problem.
the canvas is not cleared when the size is changed.
so im thinking i will add a window reload to the button, so every time
the size is changed, the page is reloaded and the canvas is cleared.

but first i will add all the other buttons making sure they work, then
i will add the window reload to each one.