made a grid of squares
now the grid is the way i want it, before it took up the whole screen.
i had a function for each button, now there is one function that takes in the
value of the button and uses that to change the size of the grid. next is a function
that changes the color of the squares when you click and drag on them