the grid can now be drawn on
so there is now a variable that is a boolean that is set to true when
the mouse is pressed and false when it is released.

this variable is used to determine whether or not to draw the grid.
i need to now add an eraser effect to the grid. and a refresh button.
from there i can add more complex things like a color picker and a rainbow effect.