the eraser button needs work
the eraser button works in the sense that all the black tiles turn into
the original color of the tile, but it doesn't work in the sense that
it doesn't disable when i want it to. once the button is clicked, it
changes to a black button with white text, but it doesn't change back
when i click it again. i want to create a color picker and once thats active
the eraser will be disabled.