there is now a color picker
the color the picker has now been implemented into the program.
also the default behavior of the web browser has been disabled. so i can
now draw on the canvas without the browser trying to do something else.
next is to disable the erase button once draw mode is activated.