the eraser function now works
i have added a draw mode and when that button is click you can now
start drawing, if the erase button is clicked the draw mode is turned off
and you can tell because the button is no longer highlighted. also i managed
to fix the problem with the eraser automatically erasing with out mousedown
being initiated. i think the only things missing now is a default canvas size
and a clear button.