i have been styling page
i have been arranging the page to look good
i havent managed to get the canvas itself to be centered but it will do for
now. i also made a range slider to change the size of the grid. i need to
setup away so the value is also displayed on the page. i also need to apply a
background image and style the header.